move functions out of createPrayerPoint and into new hook
